refactor(onboard): split migration promotion helpers
